+++ CHANGELOG.md
@@ -6,7 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.2] - 2020-01-17
 ### Added
+- Search by domain in admin panel
+### Changed
+- Advertisers' and Publishers' statistics are updated live
+### Fixed
+- Add fee to ADS exchange request
+
+## [1.5.1] - 2020-01-16
+### Changed
+- Move NowPayments info from check to depositInfo
+### Fixed
+- Https in NowPayments callback
+- Global statistics
+
+## [1.5.0] - 2020-01-15
+### Added
+- NowPayments integration
 - Edit conversions
 ### Changed
 - CPA only page rank support
@@ -342,7 +360,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v1.4.9...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v1.5.2...develop
+[1.5.2]: https://github.com/adshares/adserver/compare/v1.5.1...v1.5.2
+[1.5.1]: https://github.com/adshares/adserver/compare/v1.5.0...v1.5.1
+[1.5.0]: https://github.com/adshares/adserver/compare/v1.4.9...v1.5.0
 [1.4.9]: https://github.com/adshares/adserver/compare/v1.4.8...v1.4.9
 [1.4.8]: https://github.com/adshares/adserver/compare/v1.4.7...v1.4.8
 [1.4.7]: https://github.com/adshares/adserver/compare/v1.4.6...v1.4.7